pyproject.toml
[project]
name = "golearn-study-agent"
version = "0.1.0"
description = "AI-powered Three-Pass Study Companion using Google ADK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "google-adk>=0.1.0",
    "python-dotenv>=1.0.0",
    # FastAPI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.0.0",
    # Firebase
    "firebase-admin>=6.4.0",
    # Google AI
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",  # For testing FastAPI
]
pdf = [
    "pypdf2>=3.0.0",
    "pdfplumber>=0.9.0",
]
ppt = [
    "python-pptx>=0.6.21",
]
web = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
]
all = [
    "golearn-study-agent[dev,pdf,ppt,web]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["study_agent", "app"]
